refactor mutation helper
