+++ src/helper/mutations/thread.ts
@@ -2,10 +2,8 @@ import { gql } from 'graphql-request'
 import NProgress from 'nprogress'
 import { mutate, useSWRConfig } from 'swr'
 
-import { mutationFetch } from './helper'
-import { useAuthentication } from '@/auth/use-authentication'
+import { useMutationFetch } from './use-mutation-fetch'
 import { useEntityId } from '@/contexts/entity-id-context'
-import { useLoggedInData } from '@/contexts/logged-in-data-context'
 import {
   ThreadCreateCommentInput,
   ThreadCreateThreadInput,
@@ -14,33 +12,25 @@ import {
   ThreadSetThreadStateInput,
 } from '@/fetcher/graphql-types/operations'
 
-export function useThreadArchivedMutation() {
-  const auth = useAuthentication()
-  const entityId = useEntityId()
-  const loggedInData = useLoggedInData()
-  const { cache } = useSWRConfig()
-
-  const mutation = gql`
-    mutation threadSetArchived($input: ThreadSetThreadArchivedInput!) {
-      thread {
-        setThreadArchived(input: $input) {
-          success
-        }
+const threadArchiveMutation = gql`
+  mutation threadSetArchived($input: ThreadSetThreadArchivedInput!) {
+    thread {
+      setThreadArchived(input: $input) {
+        success
       }
     }
-  `
+  }
+`
 
-  const setThreadArchivedMutation = async function (
-    input: ThreadSetThreadArchivedInput
-  ) {
+export function useThreadArchivedMutation() {
+  const entityId = useEntityId()
+  const mutationFetch = useMutationFetch()
+  const { cache } = useSWRConfig()
+
+  return async function (input: ThreadSetThreadArchivedInput) {
     NProgress.start()
 
-    const success = await mutationFetch(
-      auth,
-      mutation,
-      input,
-      loggedInData?.strings.mutations.errors
-    )
+    const success = await mutationFetch(threadArchiveMutation, input)
 
     if (success) {
       await mutate(`comments::${entityId}`)
@@ -49,146 +39,102 @@ export function useThreadArchivedMutation() {
     }
     return success
   }
-
-  return async (input: ThreadSetThreadArchivedInput) =>
-    await setThreadArchivedMutation(input)
 }
+
+const setThreadStateMutation = gql`
+  mutation threadSetState($input: ThreadSetThreadStateInput!) {
+    thread {
+      setThreadState(input: $input) {
+        success
+      }
+    }
+  }
+`
 
 export function useSetThreadStateMutation() {
-  const auth = useAuthentication()
   const entityId = useEntityId()
-  const loggedInData = useLoggedInData()
+  const mutationFetch = useMutationFetch()
   const { cache } = useSWRConfig()
 
-  const mutation = gql`
-    mutation threadSetState($input: ThreadSetThreadStateInput!) {
-      thread {
-        setThreadState(input: $input) {
-          success
-        }
-      }
-    }
-  `
-
-  const setThreadStateMutation = async function (
-    input: ThreadSetThreadStateInput
-  ) {
-    const success = await mutationFetch(
-      auth,
-      mutation,
-      input,
-      loggedInData?.strings.mutations.errors
-    )
+  return async function (input: ThreadSetThreadStateInput) {
+    const success = await mutationFetch(setThreadStateMutation, input)
 
     if (success) await mutate(`comments::${entityId}`)
     resetAllThreadsCache(cache)
     return success
   }
-
-  return async (input: ThreadSetThreadStateInput) =>
-    await setThreadStateMutation(input)
 }
+
+const setCommentStateMutation = gql`
+  mutation threadSetCommentState($input: ThreadSetCommentStateInput!) {
+    thread {
+      setCommentState(input: $input) {
+        success
+      }
+    }
+  }
+`
 
 export function useSetCommentStateMutation() {
-  const auth = useAuthentication()
   const entityId = useEntityId()
-  const loggedInData = useLoggedInData()
+  const mutationFetch = useMutationFetch()
   const { cache } = useSWRConfig()
 
-  const mutation = gql`
-    mutation threadSetCommentState($input: ThreadSetCommentStateInput!) {
-      thread {
-        setCommentState(input: $input) {
-          success
-        }
-      }
-    }
-  `
-
-  const setCommentStateMutation = async function (
-    input: ThreadSetCommentStateInput
-  ) {
-    const success = await mutationFetch(
-      auth,
-      mutation,
-      input,
-      loggedInData?.strings.mutations.errors
-    )
+  return async function (input: ThreadSetCommentStateInput) {
+    const success = await mutationFetch(setCommentStateMutation, input)
 
     if (success) await mutate(`comments::${entityId}`)
     resetAllThreadsCache(cache)
     return success
   }
-
-  return async (input: ThreadSetCommentStateInput) =>
-    await setCommentStateMutation(input)
 }
 
-export function useCreateThreadMutation() {
-  const auth = useAuthentication()
-  const loggedInData = useLoggedInData()
-  const { cache } = useSWRConfig()
-
-  const mutation = gql`
-    mutation createThread($input: ThreadCreateThreadInput!) {
-      thread {
-        createThread(input: $input) {
-          success
-        }
+const createThreadMutation = gql`
+  mutation createThread($input: ThreadCreateThreadInput!) {
+    thread {
+      createThread(input: $input) {
+        success
       }
     }
-  `
-  const createThreadMutation = async function (input: ThreadCreateThreadInput) {
-    const success = await mutationFetch(
-      auth,
-      mutation,
-      input,
-      loggedInData?.strings.mutations.errors
-    )
+  }
+`
+
+export function useCreateThreadMutation() {
+  const mutationFetch = useMutationFetch()
+  const { cache } = useSWRConfig()
+
+  return async function (input: ThreadCreateThreadInput) {
+    const success = await mutationFetch(createThreadMutation, input)
 
     if (success) await mutate(`comments::${input.objectId}`)
     resetAllThreadsCache(cache)
     return success
   }
-
-  return async (input: ThreadCreateThreadInput) =>
-    await createThreadMutation(input)
 }
 
-export function useCreateCommentMutation() {
-  const auth = useAuthentication()
-  const entityId = useEntityId()
-  const loggedInData = useLoggedInData()
-  const { cache } = useSWRConfig()
-
-  const mutation = gql`
-    mutation createComment($input: ThreadCreateCommentInput!) {
-      thread {
-        createComment(input: $input) {
-          success
-        }
+const createCommentMutation = gql`
+  mutation createComment($input: ThreadCreateCommentInput!) {
+    thread {
+      createComment(input: $input) {
+        success
       }
     }
-  `
+  }
+`
 
-  const createCommentMutation = async function (
-    input: ThreadCreateCommentInput
-  ) {
-    const success = await mutationFetch(
-      auth,
-      mutation,
-      input,
-      loggedInData?.strings.mutations.errors
-    )
+export function useCreateCommentMutation() {
+  const mutationFetch = useMutationFetch()
+  const entityId = useEntityId()
+  const { cache } = useSWRConfig()
+
+  return async function (input: ThreadCreateCommentInput) {
+    const success = await mutationFetch(createCommentMutation, input)
 
     if (success) await mutate(`comments::${entityId}`)
     resetAllThreadsCache(cache)
 
     return success
   }
-
-  return async (input: ThreadCreateCommentInput) =>
-    await createCommentMutation(input)
 }
 
 function resetAllThreadsCache(cache: unknown) {
