+++ src/helper/mutations/use-entity-sort-mutation.ts
@@ -1,33 +1,22 @@
 import { gql } from 'graphql-request'
 
-import { mutationFetch } from './helper'
-import { useAuthentication } from '@/auth/use-authentication'
-import { useLoggedInData } from '@/contexts/logged-in-data-context'
+import { useMutationFetch } from './use-mutation-fetch'
 import { EntitySortInput } from '@/fetcher/graphql-types/operations'
 
-export function useEntitySortMutation() {
-  const auth = useAuthentication()
-  const loggedInData = useLoggedInData()
-
-  const mutation = gql`
-    mutation sort($input: EntitySortInput!) {
-      entity {
-        sort(input: $input) {
-          success
-        }
+const mutation = gql`
+  mutation sort($input: EntitySortInput!) {
+    entity {
+      sort(input: $input) {
+        success
       }
     }
-  `
-
-  const sortMutation = async function (input: EntitySortInput) {
-    const success = await mutationFetch(
-      auth,
-      mutation,
-      input,
-      loggedInData?.strings.mutations.errors
-    )
-
-    return success
   }
-  return async (input: EntitySortInput) => await sortMutation(input)
+`
+
+export function useEntitySortMutation() {
+  const mutationFetch = useMutationFetch()
+
+  return async function (input: EntitySortInput) {
+    return await mutationFetch(mutation, input)
+  }
 }
