+++ src/helper/mutations/use-user-set-description-mutation.ts
@@ -1,34 +1,28 @@
 import { gql } from 'graphql-request'
 
 import { showToastNotice } from '../show-toast-notice'
-import { mutationFetch } from './helper'
+import { useMutationFetch } from './use-mutation-fetch'
 import { useAuthentication } from '@/auth/use-authentication'
 import { useLoggedInData } from '@/contexts/logged-in-data-context'
 import { UserSetDescriptionInput } from '@/fetcher/graphql-types/operations'
 
+const mutation = gql`
+  mutation setDescription($input: UserSetDescriptionInput!) {
+    user {
+      setDescription(input: $input) {
+        success
+      }
+    }
+  }
+`
+
 export function useUserSetDescriptionMutation() {
   const auth = useAuthentication()
   const loggedInData = useLoggedInData()
+  const mutationFetch = useMutationFetch()
 
-  const mutation = gql`
-    mutation setDescription($input: UserSetDescriptionInput!) {
-      user {
-        setDescription(input: $input) {
-          success
-        }
-      }
-    }
-  `
-
-  const setDescriptionMutation = async function (input: {
-    description: string
-  }) {
-    const success = await mutationFetch(
-      auth,
-      mutation,
-      input,
-      loggedInData?.strings.mutations.errors
-    )
+  return async function (input: UserSetDescriptionInput) {
+    const success = await mutationFetch(mutation, input)
 
     if (success) {
       if (!loggedInData || !auth.current) return
@@ -38,7 +32,4 @@ export function useUserSetDescriptionMutation() {
 
     return success
   }
-
-  return async (input: UserSetDescriptionInput) =>
-    await setDescriptionMutation(input)
 }
